#5: Display and edit new fields in new structure
Added several new files to better organize the editing and display of all the La
Casa fields.  I added LC-specific program and participant profiles, as well as a
La Casa participant class and a file with helper functions to construct the
display and editing html elements.  It's still not ideal, but it's cleaner and
will hopefully be quicker to change and update.