#99, #121 and #122: Add user with program access
Alter the include/add_staff.php and extend_staff_privilege.php files to use the
new Users_Program_Access table, created in the #121 work.  Since we had the
existing bug detailed in #99, I also changed the UI so that a user with
administrative access to multiple subsites must explicitly choose a subsite to
alter.  As of this commit, creating a user and granting him/her access to one or
all programs works.  The next commit will manage editing permissions for an
existing user.

Note that I have not yet changed the UI from a select to a set of checkboxes.
We should really offer a "select multiple" option of some kind for users that
need access to more than one program but less than all.  That's the real
remaining work for issue #121.