#99, #121, #122: Link new and edits users to programs
This commit switches the UI to use checkboxes instead of select dropdowns to
choose programs.  Since the program(s) to be linked to the user will now be
passed in an array, we only need to handle one case when we add or edit program
access for users.  That case is handled in the revamped createProgramQuery()
function in core/tools/auth.php.

This doesn't handle possible duplicate rows in the Users_Program_Access table.
Changes on this branch should fix both #99 and #122, since we are now requiring
a user with administrative access to multiple subsites to explicitly choose one
subsite to edit.