#133: Display human-readable values
Some fields are saved via a select menu, which saves integer or character values
instead of the human-readable names for those values.  This commit uses the
existing arrays of possible values to display the human-readable value instead
of the one stored in the database.

Note that while doing this I opened #134, because the current arrays for majors
and minors will change every time a new major or minor is created, which would
have changed the displayed values of currently entered majors and minors!  That
won't do.  I addressed it for now by making the array keys equal to the values,
but instead of keeping it that way I'll make a major table that stores all
majors with a unique identifier.