#117: Add SQL file to show funders and schedule correctly.
Running this SQL file is necessary to make the program dates show up correctly.
Without it, mysql will complain that the Funder_ID is not a column in
Subcategory_Dates and the dates will not show up.

Update the sample data with the Funder_ID column and row in Institution_Types
accordingly.