#121: Correct Enlace reports with new program access check.
Use the new Users_Program_Access table to govern the results in Enlace reports.
Remove the old "if has access to all programs" checks in favor of a system that
works no matter how many programs the logged-in user has access to.  Also,
return the createProgramQuery() function to auth.php (I think it was
accidentally removed in a merge).