#127: Save email alert fields as dates
La Casa includes some fields that represent alerts about emails being sent to
residents.  Previously, we had saved those as yes/no indications of whether an
email had been sent or not.  Instead, these need to be saved as dates to
indicate when emails were sent.  Make that change in the front end.  Update the
datepicker to show reasonable years for those emails.  Add a script to alter the
types of those columns and update the sample data accordingly.