#121: Adjust exports to show results from multiple programs
Since we've allowed users to have access to multiple programs (rather than one
or all), that change needed to be reflected in the exports.  I adjusted the
queries to use the set of programs that the user has access to as an OR group,
so they'll get any results related to any of those programs.