#141: Start altering fields and updating UI
Remove two fields that are calculated, not entered.  Show reasons for
independence as checkboxes instead of individual dropdowns.

Note that we have a two masters problem with La Casa data entry.  The same
information is being entered both on the dashboard page and on the lc_profile.
These fields should only be generated in one place, so that we only have to
update them in one place.