#141: Add new move in and out fields.
Replace the move in and out dates with seasons and years.  I won't be replacing
the database fields, since we already completed the import with dates into those
individual fields.  Because of that, I'll continue displaying the move in and
move out dates, but they won't be editable.  Future participants will be entered
with move in and out seasons and years, not dates.