#142: Add new values for several fields.
Colleges, school years, majors, and minors are all displayed as select
elements constructed from existing values in the database, but users
will occasionally want to add new values to those arrays.  This commit
allows them to do so.  We save colleges, cohorts, and statuses in their
own tables and link them to residents by a foreign key.  We save school
years, majors, and minors as strings in the LC_Basics table, but show
them as a select element to avoid duplicates like 2013-14 and 2013-2014
as much as possible.  I created a new function to build a select element
and an input element for those cases where we need both.