#144: Loans and financial information now editable
This change didn't strictly need to include a database change, but I
moved the loans information to the "LC_Basics" table since that more
accurately represents the way we're showing the information to users.
In the future I think they will want to show loans over time, but that
isn't really being captured in the existing information.