Fix javascript errors for #156
Two similar but unrelated errors were conspiring here: the first was
that "failAlert" was undefined in the top-level page, participants.php.
That function shouldn't have been called in the first place, since the
admin-level user had the right to delete the participants.  It was only
being called because of an unrelated javascript error in
search_participants.php.  There, the .fail() call was misplaced after
the end of a conditional, instead of at the end of the post() call.

That fail() call will return the "failAlert()" anyway, so I removed the
original (duplicate) call to failAlert in participants.php.