Account for dropped attendees: #160
I was seeing some negative dosage hours in this report, and I think that
was due to the fact that I wasn't accounting for people who dropped out
of the program.  They shouldn't be counted after dropping, nor should
any absences of theirs be counted after they drop out of the program.

Also, manage SQL escaping, ahem.

Currently people who have dropped are not counted towards enrollment or
dosage, because of the way the latter is calculated (i.e., with
person-days provided).  The number of people used in that calculation is
the number of enrollees.  Those who've dropped aren't counted.

The right thing to do is to count the people who've dropped for the days
before they drop.  That will add some complexity to the calculation, but
would make it actually correct.