Move participant deletion to avoid constraints
Because we don't have "ON DELETE CASCADE" set in all database tables,
deleting a participant failed when the participant had certain
relationships.  This commit moves the participant deletion call so that
it occurs after all participant relationships have been removed, but
this is just one instance of a larger problem discussed in issue #169.
The bigger fix is to implement ON DELETE CASCADE for all foreign key
relationships (or at least to have a discussion about that), and to add
error handling to database calls, so that they don't fail silently from
a user perspective.