Display error for invalid dates, update help text: #170
Move javascript out of the main HTML/PHP content and up to the top of
the file.  Point to it from the onclick.  Split it into two functions.
Remove the ajax loader even that isn't directly related to this issue.
Update helptext and add warning.  Add new class ("error") for warnings.
Set datepicker format explicitly to yyyy-mm-dd.