Fix more instances of limiting to 2016
A user noted that date menus in the app all stop at 2016.  We are now in
2017, so that needs to be extended.  I started this in 7def088.  Note in
this commit that at least one input still has manual years (the "consent
forms" dropdown).  This is obviously not good, and will result in this
problem showing up again in a few years, if anyone is still using this
system.