Fix mentorship export: #185
The query for exporting mentorship hours incorrectly referred to
Programs instead of Sessions.  The export appeared to be working as long
as the number of programs and sessions was roughly equivalent, though
the "program" named in the export was never correct.  The illusion of
success was caused by the fact that if there was a session with id "10,"
there was likely also a program with id "10."  Now that there are
hundreds of sessions and only dozens of programs, it's highly likely
that a session ID will not correspond to any program.  Because this was
an inner join, mentorship hours without a corresponding program ID were
not returned by the query.

Note that this is one instance of a general problem with the database
schema: programs and sessions are unclearly differentiated.  In some
cases a table refers to a "program id" when it really means a "session
id."  The converse is less frequent.