Fix #221: Enlace: Add survey entry point link
This adds an entry point to the intake/impact surveys without logging
in.  To do this, it creates an effectively unique id based on the name
of the person, the session, and a salt.  This was chosen over
maintaining a GUID list somewhere in the database for ease of
development and maintenance.