Issue #222: Add day and start/end to session
Session needs to have their own Days of Week, and Daily Start / Daily
End apart from the program, though we keep the program ones for
defaults.  This adds those fields to the database as well as updating
the UI for them.

Issue #222: Enlace: Sessions have their own Days of week and Daily Start
            and End hours