Fix #244: Enlace: Require new session information
This is only a client side fix for new sessions.  Because the database
is not enforcing it, and because the backend, it's technically possible
to get around this (even by just editing one).  But "require" here
really just means "strongly suggest," and so the code allows workarounds
in the case someone REALLY wants to.